pipeline/pyproject.toml
[project]
name = "larp-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "David Lin-Clark", email = "david@woolymango.com" }
]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.95.0",
    "faster-whisper>=1.2.1",
    "numpy>=2.4.4",
    "onnxruntime<1.20",
    "pillow>=12.2.0",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "typer>=0.24.1",
]

[project.scripts]
larp-pipeline = "larp_pipeline:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
